[#367] Refactor modules/signup to modules/stakeholder-signup & refs
Thus we'll need differenciate between StakeholderSignup and EntitySignup
