+++ frontend/src/modules/stakeholder-signup/styles.scss
