+++ frontend/src/modules/stakeholder-signup/view-style.scss
