+++ frontend/src/root.jsx
@@ -13,13 +13,13 @@ import Landing from "./modules/landing/view";
 import Browse from "./modules/browse/view";
 import AddEvent from "./modules/events/view";
 import logo from "./images/GPML-temporary-logo-horiz.jpg";
-import StakeholderSignupModal from "./modules/signup/signup-modal";
+import StakeholderSignupModal from "./modules/stakeholder-signup/signup-modal";
 import ModalWarningUser from "./utils/modal-warning-user";
 import api from "./utils/api";
 import { storage } from "./utils/storage";
 import { UIStore } from "./store.js";
 import ProfileView from "./modules/profile/view";
-import SignupView from "./modules/signup/view";
+import StakeholderSignupView from "./modules/stakeholder-signup/view";
 import DetailsView from "./modules/details/view";
 import Footer from "./footer";
 import uniqBy from "lodash/uniqBy";
@@ -178,7 +178,7 @@ const Root = () => {
             </div>
             {!isAuthenticated ? (
               <div className="rightside">
-                <Link to="/signup">Join the GPML</Link>
+                <Link to="/stakeholder-signup">Join the GPML</Link>
                 &nbsp;&nbsp;|&nbsp;&nbsp;
                 <Link to="/" onClick={loginWithPopup}>
                   Sign in
@@ -335,7 +335,7 @@ const Root = () => {
           path="/profile"
           render={(props) => <ProfileView {...{ ...props }} />}
         />
-        <Route path="/signup" render={(props) => <SignupView {...props} />} />
+        <Route path="/stakeholder-signup" render={(props) => <StakeholderSignupView {...props} />} />
         <Route
           path="/:type(project|action_plan|policy|technical_resource|financing_resource|technology|event|organisation|stakeholder)/:id"
           render={(props) => (
